bc-ml-trading/bc-ml-trading/requirements.txt
ta
pandas
pyzmq
dotenv